+++ src/components/IconArrow/IconArrow.example.js
@@ -1,0 +1,17 @@
+import IconArrow from './IconArrow';
+
+export const IconArrowUp = {
+  component: IconArrow,
+  props: {
+    direction: 'up',
+  },
+  group: 'icons',
+};
+
+export const IconArrowDown = {
+  component: IconArrow,
+  props: {
+    direction: 'down',
+  },
+  group: 'icons',
+};